fix(operation-name): get graphql operation name from context.
previously it was reporting the first operation name in the query even the running query is different.

The bug was something like this:

```graphql
query A {
   __typename
}
query B {
   __typename
}
query C {
   __typename
}
```
Even if I run the operation `C`, it was always reporting `A` as operation name.

This PR fixes the bug :rocket: